requirements.txt
streamlit==1.29.0
langgraph==0.2.16
langchain==0.2.16
langchain-openai==0.1.25
requests==2.31.0
Pillow==10.1.0
python-dotenv==1.0.0